move composable to common module
